Add template content for section 1
I deecided to make the landing page about my portfolio.

The nav bar is for navigating to my large projects, small projects, and contact me page.

Below that is the current note worthy  project i'm working on and a button to preview the prototype, with an image of the project concept. or my face(haven't decided yet.